Implement packet number wraparound
This commit integrates a feature contributed by Maxime Hadjinlian to support
the transfer of files of arbitrary size by looping the DATA packet number. This
is a commonly implemented feature in TFTP servers, known as packet number
wraparound.

When the packet number reaches the maximum allowed by the TFTP protocol, 2^16-1
(65535), the packet number is reset to 1 to allow the transfer to continue. The
target client must of course support the feature as well.

In both pTFTP and pTFTPd, this behavior is disabled when the strict RFC1350
mode of operation is used (--rfc1350).