Check for netifaces availability
Since we are using 'distutils', it won't install dependency for us as we
could do with 'setuptools'.

It's not usefull to switch to 'setuptools' since, in order to install
'netifaces' it will try build an extension which requires to have
'python-dev' package, which is really not user-friendly.

Instead simply show an error message asking the user to install missing
dependency during setup.